backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
flask-socketio==5.3.6
python-socketio==5.8.0
numpy==1.21.2
pandas==1.3.2
scikit-learn==0.24.2
joblib==1.1.0
python-dotenv==0.19.0
